add basic raid event
