+++ Content.Server/_RF/GameTicking/Rules/MigrationRuleComponent.cs
@@ -1,10 +1,11 @@
+using Content.Server._RF.NPC;
 using Content.Shared.Destructible.Thresholds;
 using Robust.Shared.Prototypes;
 
-namespace Content.Shared._RF.GameTicking.Rules;
+namespace Content.Server._RF.GameTicking.Rules;
 
 /// <summary>
-/// Events of refugee migration to the settlement
+/// Entity migration event to the player's settlement
 /// </summary>
 [RegisterComponent]
 public sealed partial class MigrationRuleComponent : Component
@@ -26,4 +27,19 @@ public sealed partial class MigrationRuleComponent : Component
     /// </summary>
     [DataField]
     public bool AddToPops { get; set; }
+
+    /// <summary>
+    /// Minimum radius from the player's settlements where the event will take place
+    /// </summary>
+    [DataField]
+    public int RadiusFromSettlement = 30;
+
+    /// <summary>
+    /// NPC task that will be given to entities
+    /// </summary>
+    /// <remarks>
+    /// The coordinates of one of the player's settlements will be assigned as the task target
+    /// </remarks>
+    [DataField]
+    public ProtoId<NpcTaskPrototype>? Task { get; set; }
 }
