+++ Resources/Locale/en-US/_RF/game-ticking/game-presets/rim-fortress-commands.ftl
@@ -1,0 +1,7 @@
+cmd-startworldrule-decs = Starts a local world event for the player
+cmd-startworldrule-help = Starts a local world event for the player
+cmd-startworldrule-prototype = world event prototype
+cmd-startworldrule-no-settlements = The player does not have any settlements to start the event
+
+cmd-startworldrulenow-decs = Starts a local world event for the player ignoring the delay
+cmd-startworldrulenow-help = Starts a local world event for the player ignoring the delay
