+++ Resources/Prototypes/_RF/GameRules/events.yml
@@ -28,6 +28,9 @@
       - id: SpiderMigration
       - id: OreCrabMigration
 
+      # Raids
+      - id: Raid
+
 - type: entity
   id: BaseMigrationRule
   abstract: true
@@ -302,3 +305,21 @@
       amount:
         min: 3
         max: 6
+
+# Raids
+- type: entity
+  id: Raid
+  components:
+    - type: WorldRule
+      delay:
+        min: 300 #5 min
+        max: 900 #15 min
+      startOffset: 900 #15 min
+      threshold: 0
+    - type: MigrationRule
+      spawn:
+        - Raider
+      amount:
+        min: 1
+        max: 3
+      task: Raid
