+++ Resources/Prototypes/_RF/GameRules/roundstart.yml
@@ -7,9 +7,6 @@
       defaultPopsJob: Passenger
       worldEvents: !type:NestedSelector
         tableId: BasicWorldRulesTable
-      minMaxEventTiming:
-        min: 600 #10 min
-        max: 1800 #20 min
       dayDuration: 1800 #30 min
       popsComponentsOverride:
         - type: HTN
