+++ Resources/Prototypes/_RF/NPCs/raid.yml
@@ -1,0 +1,18 @@
+- type: htnCompound
+  id: RaidCompound
+  branches:
+    - tasks:
+        - !type:HTNPrimitiveTask
+          operator: !type:UtilityOperator
+            proto: NearbyMeleeTargets
+        - !type:HTNCompoundTask
+          task: AttackTargetCompound
+
+    - tasks:
+        - !type:HTNPrimitiveTask
+          operator: !type:PickNearestFreeOperator
+        - !type:HTNPrimitiveTask
+          operator: !type:MoveToOperator
+            pathfindInPlanning: true
+            stopOnLineOfSight: true
+            rangeKey: MeleeRange
