+++ Resources/Prototypes/ai_factions.yml
@@ -123,7 +123,10 @@
   - Zombie
   - Revolutionary
   - Wizard
-  - RimFortress # RimFortress
+  # RimFortress Start
+  - RimFortress
+  - Raiders
+  # RimFortress End
 
 - type: npcFaction
   id: Wizard
